Further adjust the boundary between .NET and the generic generator
