+++ src/dot_net/index.js
@@ -1,4 +1,4 @@
-import Generator from '../generator';
+import Generator from "../generator";
 
 const DATA_MODEL_DOCS_URL_PREFIX = "https://developer.openactive.io/data-model/types/";
 
@@ -20,7 +20,9 @@ class DotNet extends Generator {
     let inherits = this.calculateInherits(model.subClassOf, derivedFrom, model);
 
     // Note hasBaseClass is used here to ensure that assumptions about schema.org fields requiring overrides are not applied if the base class doesn't exist in the model
-    let hasBaseClass = inherits != "Schema.NET.JsonLdObject";
+    let hasBaseClass = inherits !== "Schema.NET.JsonLdObject";
+
+    let doc = this.createModelDoc(model, models);
 
     return `
 using Newtonsoft.Json;
@@ -31,14 +33,7 @@ using System.Runtime.Serialization;
 namespace OpenActive.NET
 {
     /// <summary>
-    /// ${this.getPropNameFromFQP(model.type) != model.type
-      ? `[NOTICE: This is a beta class, and is highly likely to change in future versions of this library.]. `
-      : ""}${this.createCommentFromDescription(model.description).
-      replace(/\n/g, "\n    /// ")}
-    /// ${derivedFrom ? `This type is derived from [` + derivedFromName + `](` +
-      derivedFrom + `)` + (derivedFrom.indexOf("schema.org") > -1
-        ? ", which means that any of this type's properties within schema.org may also be used. Note however the properties on this page must be used in preference if a relevant property is available"
-        : "") + "." : ""}
+    /// ${doc.join(`\n    /// `)}
     /// </summary>
     [DataContract]
     public partial class ${this.convertToCamelCase(
@@ -222,57 +217,29 @@ namespace OpenActive.NET
     return types.length > 1 ? `SingleValues<${types.join(", ")}>` : types[0];
   }
 
-  isArray (prop) {
-    return prop.indexOf("ArrayOf") == 0;
-  }
 
-  getPropLinkFromFQP (prop) {
-    if (prop.lastIndexOf("/") > -1) {
-      return prop.replace("ArrayOf#", "");
-    } else if (prop.lastIndexOf("#") > -1) {
-      return DATA_MODEL_DOCS_URL_PREFIX +
-        prop.substring(prop.lastIndexOf("#") + 1).toLowerCase();
-    } else return "#";
-  }
-
-  getPropNameFromFQP (prop) {
-    if (prop === null || prop === undefined) return null;
-    //Just the characters after the last /, # or :
-    let match = prop.match(/[/#:]/g);
-    let lastIndex = match === null ? -1 : prop.lastIndexOf(
-      match[match.length - 1]);
-    if (lastIndex > -1) {
-      return prop.substring(lastIndex + 1);
-    } else return prop;
-  }
-
-  createDescriptionWithExample (field) {
-    if (field.requiredContent) {
-      return "Must always be present and set to " +
-        this.renderCode(field.requiredContent, field.fieldName,
-          field.requiredType);
+  calculateInherits (subClassOf, derivedFrom, model) {
+    // Prioritise subClassOf over derivedFrom
+    if (subClassOf) {
+      let subClassOfName = this.convertToCamelCase(
+        this.getPropNameFromFQP(subClassOf));
+      if (this.includedInSchema(subClassOf)) {
+        return `Schema.NET.${subClassOfName}`;
+      } else {
+        return `${subClassOfName}`;
+      }
+    } else if (derivedFrom) {
+      let derivedFromName = this.convertToCamelCase(
+        this.getPropNameFromFQP(derivedFrom));
+      if (this.includedInSchema(derivedFrom)) {
+        return `Schema.NET.${derivedFromName}`;
+      } else {
+        // Note if derived from is outside of schema.org there won't be a base class, but it will still be JSON-LD
+        return `Schema.NET.JsonLdObject`;
+      }
     } else {
-      let betaWarning = field.extensionPrefix == "beta"
-        ? "[NOTICE: This is a beta field, and is highly likely to change in future versions of this library.] \n"
-        : "";
-      return `<summary>\n${betaWarning}${field.description.join(
-        " \n")}\n</summary>`
-        + (field.example ? "\n<example>\n" +
-          this.renderCode(field.example, field.fieldName, field.requiredType) +
-          "\n</example>" : "");
-    }
-  }
-
-  renderCode (code, fieldName, requiredType) {
-    if (typeof code === "object") {
-      return "<code>\n" + (fieldName ? `"` + fieldName + `": ` : "") +
-        JSON.stringify(code, null, 2) + "\n</code>";
-    } else {
-      let isNumber = requiredType &&
-        (requiredType.indexOf("Integer") > -1 || requiredType.indexOf("Float") >
-          -1);
-      return "<code>\n" + (fieldName ? `"` + fieldName + `": ` : "") +
-        (isNumber ? code : `"` + code + `"`) + "\n</code>";
+      // In the model everything is one or the other (at a minimum must inherit https://schema.org/Thing)
+      throw new Error("No base class specified for: " + model.type);
     }
   }
 }
