+++ src/generator.js
@@ -255,31 +255,6 @@ class Generator {
     return augFields;
   }
 
-  calculateInherits (subClassOf, derivedFrom, model) {
-    // Prioritise subClassOf over derivedFrom
-    if (subClassOf) {
-      let subClassOfName = this.convertToCamelCase(
-        this.getPropNameFromFQP(subClassOf));
-      if (this.includedInSchema(subClassOf)) {
-        return `Schema.NET.${subClassOfName}`;
-      } else {
-        return `${subClassOfName}`;
-      }
-    } else if (derivedFrom) {
-      let derivedFromName = this.convertToCamelCase(
-        this.getPropNameFromFQP(derivedFrom));
-      if (this.includedInSchema(derivedFrom)) {
-        return `Schema.NET.${derivedFromName}`;
-      } else {
-        // Note if derived from is outside of schema.org there won't be a base class, but it will still be JSON-LD
-        return `Schema.NET.JsonLdObject`;
-      }
-    } else {
-      // In the model everything is one or the other (at a minimum must inherit https://schema.org/Thing)
-      throw new Error("No base class specified for: " + model.type);
-    }
-  }
-
   compareFields (xField, yField) {
     let x = xField.fieldName.toLowerCase();
     let y = yField.fieldName.toLowerCase();
@@ -388,6 +363,90 @@ class Generator {
   includedInSchema (url) {
     if (!url) return false;
     return url.indexOf("//schema.org") > -1 || url.indexOf("schema:") == 0;
+  }
+
+  createModelDoc (model, models) {
+    let derivedFrom = this.getPropertyWithInheritance("derivedFrom", model,
+      models);
+    let derivedFromName = this.convertToCamelCase(
+      this.getPropNameFromFQP(derivedFrom));
+
+    let docLines = [
+      (this.getPropNameFromFQP(model.type) !== model.type) &&
+      `[NOTICE: This is a beta class, and is highly likely to change in future versions of this library.].`,
+      this.createCommentFromDescription(model.description),
+    ];
+
+    if (derivedFrom) {
+      let text = `This type is derived from [${derivedFromName}](${derivedFrom})`;
+
+      if (derivedFrom.indexOf("schema.org")) {
+        text += ", which means that any of this type's properties within schema.org may also be used. Note however the properties on this page must be used in preference if a relevant property is available";
+      }
+
+      text += ".";
+
+      docLines.push(text);
+    }
+
+    return docLines.
+      filter((val) => val).
+      reduce((acc, val) => acc.concat(val.split("\n")), []);
+  }
+
+
+  isArray (prop) {
+    return prop.indexOf("ArrayOf") == 0;
+  }
+
+  getPropLinkFromFQP (prop) {
+    if (prop.lastIndexOf("/") > -1) {
+      return prop.replace("ArrayOf#", "");
+    } else if (prop.lastIndexOf("#") > -1) {
+      return DATA_MODEL_DOCS_URL_PREFIX +
+        prop.substring(prop.lastIndexOf("#") + 1).toLowerCase();
+    } else return "#";
+  }
+
+  getPropNameFromFQP (prop) {
+    if (prop === null || prop === undefined) return null;
+    //Just the characters after the last /, # or :
+    let match = prop.match(/[/#:]/g);
+    let lastIndex = match === null ? -1 : prop.lastIndexOf(
+      match[match.length - 1]);
+    if (lastIndex > -1) {
+      return prop.substring(lastIndex + 1);
+    } else return prop;
+  }
+
+  createDescriptionWithExample (field) {
+    if (field.requiredContent) {
+      return "Must always be present and set to " +
+        this.renderCode(field.requiredContent, field.fieldName,
+          field.requiredType);
+    } else {
+      let betaWarning = field.extensionPrefix == "beta"
+        ? "[NOTICE: This is a beta field, and is highly likely to change in future versions of this library.] \n"
+        : "";
+      return `<summary>\n${betaWarning}${field.description.join(
+        " \n")}\n</summary>`
+        + (field.example ? "\n<example>\n" +
+          this.renderCode(field.example, field.fieldName, field.requiredType) +
+          "\n</example>" : "");
+    }
+  }
+
+  renderCode (code, fieldName, requiredType) {
+    if (typeof code === "object") {
+      return "<code>\n" + (fieldName ? `"` + fieldName + `": ` : "") +
+        JSON.stringify(code, null, 2) + "\n</code>";
+    } else {
+      let isNumber = requiredType &&
+        (requiredType.indexOf("Integer") > -1 || requiredType.indexOf("Float") >
+          -1);
+      return "<code>\n" + (fieldName ? `"` + fieldName + `": ` : "") +
+        (isNumber ? code : `"` + code + `"`) + "\n</code>";
+    }
   }
 }
